Data-Validation-Agent/requirements.txt
google-adk
google-cloud-bigquery
requests
fastapi
uvicorn